// src/components/Forms/TimeTableForm/types.d.ts
export type time_table = {
  Mon?: [string?, string?];
  Tue?: [string?, string?];
  Wed?: [string?, string?];
  Thu?: [string?, string?];
  Fri?: [string?, string?];
  Sat?: [string?, string?];
  Sun?: [string?, string?];
};

export type radioGroup = "Круглосуточно" | "Настроить" | string;